#91 mysql/sqlite의 getNextSequence()에 대해 10000개 단위로 삭제하도록하고 mysql에서 0이라는 값을 입력하도록 수정
git-svn-id: http://xe-core.googlecode.com/svn/sandbox@2364 201d5d3c-b55e-5fd7-737f-ddc643e51545

+++ classes/db/DBMysql.class.php
@@ -177,11 +177,13 @@
          * @brief 1씩 증가되는 sequence값을 return (mysql의 auto_increment는 sequence테이블에서만 사용)
          **/
         function getNextSequence() {
-            $query = sprintf("insert into `%ssequence` (seq) values ('')", $this->prefix);
+            $query = sprintf("insert into `%ssequence` (seq) values ('0')", $this->prefix);
             $this->_query($query);
             $sequence = mysql_insert_id();
-            $query = sprintf("delete from  `%ssequence` where seq < %d", $this->prefix, $sequence);
-            $this->_query($query);
+            if($seqnece % 10000 == 0) {
+              $query = sprintf("delete from  `%ssequence` where seq < %d", $this->prefix, $sequence);
+              $this->_query($query);
+            }
 
             return $sequence;
         }
